feat: add esm.mjs
